Reorganize file content, make symbols internal
